feat(runtime-c-api) Extract the `table` module.
